Fix to the algorithm inserting a win pre block.
It was possible for an opponent who knew the strategy of this program
to beat the algorithm assuming they play first and play corner then
opposite corner. Following my move to the side, they need to not block
and instead play the corner opposite of the computer's move. This sets
up a scenario where the algorithm plays a block in the first square it
comes to instead of winning. This lets the opponent win.

The logic now wins if possible before blocking.